Factor out QuoteEdit from show view
Show view uses QuoteEdit component for owned Quote
